requirements.txt
haystack-ai
milvus-haystack
jina-haystack